Migrate tests related to special types
