+++ src/Http/Http.Extensions/test/RequestDelegateFactoryTests.cs
@@ -1188,111 +1188,6 @@ public partial class RequestDelegateFactoryTests : LoggedTest
         Assert.Same(myOriginalService, httpContext.Items["service"]);
     }
 
-    [Fact]
-    public async Task RequestDelegatePopulatesHttpContextParameterWithoutAttribute()
-    {
-        HttpContext? httpContextArgument = null;
-
-        void TestAction(HttpContext httpContext)
-        {
-            httpContextArgument = httpContext;
-        }
-
-        var httpContext = CreateHttpContext();
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Same(httpContext, httpContextArgument);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePassHttpContextRequestAbortedAsCancellationToken()
-    {
-        CancellationToken? cancellationTokenArgument = null;
-
-        void TestAction(CancellationToken cancellationToken)
-        {
-            cancellationTokenArgument = cancellationToken;
-        }
-
-        using var cts = new CancellationTokenSource();
-        var httpContext = CreateHttpContext();
-        // Reset back to default HttpRequestLifetimeFeature that implements a setter for RequestAborted.
-        httpContext.Features.Set<IHttpRequestLifetimeFeature>(new HttpRequestLifetimeFeature());
-        httpContext.RequestAborted = cts.Token;
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.RequestAborted, cancellationTokenArgument);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePassHttpContextUserAsClaimsPrincipal()
-    {
-        ClaimsPrincipal? userArgument = null;
-
-        void TestAction(ClaimsPrincipal user)
-        {
-            userArgument = user;
-        }
-
-        var httpContext = CreateHttpContext();
-        httpContext.User = new ClaimsPrincipal();
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.User, userArgument);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePassHttpContextRequestAsHttpRequest()
-    {
-        HttpRequest? httpRequestArgument = null;
-
-        void TestAction(HttpRequest httpRequest)
-        {
-            httpRequestArgument = httpRequest;
-        }
-
-        var httpContext = CreateHttpContext();
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Request, httpRequestArgument);
-    }
-
-    [Fact]
-    public async Task RequestDelegatePassesHttpContextRresponseAsHttpResponse()
-    {
-        HttpResponse? httpResponseArgument = null;
-
-        void TestAction(HttpResponse httpResponse)
-        {
-            httpResponseArgument = httpResponse;
-        }
-
-        var httpContext = CreateHttpContext();
-
-        var factoryResult = RequestDelegateFactory.Create(TestAction);
-        var requestDelegate = factoryResult.RequestDelegate;
-
-        await requestDelegate(httpContext);
-
-        Assert.Equal(httpContext.Response, httpResponseArgument);
-    }
-
     public static IEnumerable<object[]> ChildResult
     {
         get
